Abstract out admin handlers into admin controller
